wip: agent checkpoint 11:56
